Rename Iterator to Memory Access Strategy
- rename was necessary to avoid confusions with fancy iterator
- rename and flat namespace of the mem access strategy
- move related files to new folder
