+++ include/vikunja/reduce/reduce.hpp
@@ -1,4 +1,4 @@
-/* Copyright 2021 Hauke Mewes, Simeon Ehrig
+/* Copyright 2022 Hauke Mewes, Simeon Ehrig
  *
  * This file is part of vikunja.
  *
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include <vikunja/mem/iterator/PolicyBasedBlockIterator.hpp>
+#include <vikunja/access/PolicyBasedBlockStrategy.hpp>
 #include <vikunja/operators/operators.hpp>
 #include <vikunja/reduce/detail/BlockThreadReduceKernel.hpp>
 #include <vikunja/reduce/detail/SmallProblemReduceKernel.hpp>
@@ -54,7 +54,7 @@ namespace vikunja
          * @tparam WorkDivPolicy The working division policy. Defaults to a templated value depending on the
          * accelerator. For the API of this, see workdiv/BlockBasedWorkDiv.hpp
          * @tparam MemAccessPolicy The memory access policy. Defaults to a templated value depending on the
-         * accelerator. For the API of this, see mem/iterator/PolicyBasedBlockIterator
+         * accelerator. For the API of this, see vikunja::MemAccess::PolicyBasedBlockStrategy
          * @tparam TTransformFunc Type of the transform operator.
          * @tparam TReduceFunc Type of the reduce operator.
          * @tparam TInputIterator Type of the input iterator. Should be a pointer-like type.
@@ -77,7 +77,7 @@ namespace vikunja
         template<
             typename TAcc,
             typename WorkDivPolicy = vikunja::workdiv::BlockBasedPolicy<TAcc>,
-            typename MemAccessPolicy = vikunja::mem::iterator::MemAccessPolicy<TAcc>,
+            typename MemAccessPolicy = vikunja::MemAccess::MemAccessPolicy<TAcc>,
             typename TTransformFunc,
             typename TReduceFunc,
             typename TInputIterator,
@@ -219,7 +219,7 @@ namespace vikunja
          * @tparam WorkDivPolicy The working division policy. Defaults to a templated value depending on the
          * accelerator. For the API of this, see workdiv/BlockBasedWorkDiv.hpp
          * @tparam MemAccessPolicy The memory access policy. Defaults to a templated value depending on the
-         * accelerator. For the API of this, see mem/iterator/PolicyBasedBlockIterator
+         * accelerator. For the API of this, see vikunja::MemAccess::PolicyBasedBlockStrategy
          * @tparam TTransformFunc Type of the transform operator.
          * @tparam TReduceFunc Type of the reduce operator.
          * @tparam TInputIterator Type of the input iterator. Should be a pointer-like type.
@@ -241,7 +241,7 @@ namespace vikunja
         template<
             typename TAcc,
             typename WorkDivPolicy = vikunja::workdiv::BlockBasedPolicy<TAcc>,
-            typename MemAccessPolicy = vikunja::mem::iterator::MemAccessPolicy<TAcc>,
+            typename MemAccessPolicy = vikunja::MemAccess::MemAccessPolicy<TAcc>,
             typename TTransformFunc,
             typename TReduceFunc,
             typename TInputIterator,
@@ -293,7 +293,7 @@ namespace vikunja
         template<
             typename TAcc,
             typename WorkDivPolicy = vikunja::workdiv::BlockBasedPolicy<TAcc>,
-            typename MemAccessPolicy = vikunja::mem::iterator::MemAccessPolicy<TAcc>,
+            typename MemAccessPolicy = vikunja::MemAccess::MemAccessPolicy<TAcc>,
             typename TFunc,
             typename TInputIterator,
             typename TDevAcc,
@@ -352,7 +352,7 @@ namespace vikunja
         template<
             typename TAcc,
             typename WorkDivPolicy = vikunja::workdiv::BlockBasedPolicy<TAcc>,
-            typename MemAccessPolicy = vikunja::mem::iterator::MemAccessPolicy<TAcc>,
+            typename MemAccessPolicy = vikunja::MemAccess::MemAccessPolicy<TAcc>,
             typename TFunc,
             typename TInputIterator,
             typename TDevAcc,
